Update patch set 3
Patch Set 3:

> The code in quick says that it is misplaced (with a TODO to move to
 > compiler driver). Considering TODO-ing rather than replicating the
 > wrong approach?

Like I said, I tried to move it in a shared place (like IsPathologicalCase), but the logic is not so easy to extract. Also, if we want to do the same, it includes checks on a MirGraph and overrides previous decisions. So we'd need to find some kind of abstract computation of, e.g., fieldget_rations, arithmetic_ratios, etc.

But I can add a TODO in following changes.

Patch-set: 3