Update patch set 8
Patch Set 8:

So yes, I agree that the overloaded meaning on the MIRs is not good design. But I also don't want this information in a side data structure. My big concern is that your changes here are going to cause some incompatibilities on our side :)

I would prefer still to reflect suspend logic explicitly in the graph. Either via the marking of instructions like we currently do or via a new extended instruction "kSuspendCheck". Basically we want it for correctness to not be able to move across barriers and it is better for barrier to be in graph instead of side structure.

So in my mind, to continue with current design, let's also make sure that correct flags are on every branching instructions, which will include the fused comparison branching.

Patch-set: 8
Reviewer: Razvan A Lupusoru <1013855@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0