Update patch set 1
Patch Set 1:

> How about a compromise? I took another look at our use case.  It
 > turns out that the only difference is that we also want to turn on
 > store_method_addr for FLOAT uses as well as double.
 > 
 > Can you:
 > - Add the float cases too (Add the cases in AnalyzeMIR, and enhance
 > AnalyseFPInstruction to handle non-wide cases.
 > 
 > or
 > - Make all the AnalyzeMIR virtual (so that I can extend as needed)

So I'm not sure what contract we're trying to make "Analyze.." have. Is the issue with this change that you have more use-cases? If so why not upstream a patch to add them, why the need for more virtuals? I'd prefer a name like "NeedsCodeBasePtr" as being more intention revealing than "Analyze X86". Why in the base case are we treating floats and doubles differently it seems a bug no? As I don't have access to your change it seems the burden should be on you to do the work to make this work with your change, as I have no means to test that.

Patch-set: 1