Update patch set 2
Patch Set 2:

> (10 comments)
 > 
 > There are a couple of changes unrelated to inlining. 

Which ones? They are all triggered by inlining requirements. Sure I could do a change for the HGraph::InlineInto, a refactoring of TryBuildingSsa, a renaming of the HInvokeStatic to HInvokeStaticOrDirect, but I feel they are so small and part of the inlining logic, that they would look superficial standalone. Maybe you have one other example in mind?

> I'm ok with
 > this since I know how convenient it is! Need to mention though that
 > I find it easier to review when the CLs are focused :). Also, the
 > commit message becomes more meaningful as it states exactly what
 > you did and doesn't leave things out.

Patch-set: 2
Label: Verified=0