Update patch set 10
Patch Set 10:

(10 comments)

As discussed with Nicolas, Checker now expects the C1visualizer format so that the same format is used in testing and debugging. However, due to constraints of the C1 format, pattern matching isn't sufficient to recognise the end of a graph dump and hence Checker needs to partially parse the output.

Patch-set: 10
Label: Verified=0