Update patch set 1
Patch Set 1: Code-Review+1

It is somewhat surprising that IsPowerOfTwo(0) returns true as zero is not a power of two. One may argue that either the name of the function is misleading (it should rather be IsShiftedBit) or the function should return false. It is unclear to me, however, whether other parts of the code currently rely on this behaviour. So it may be a good idea to just stick with what you did. Similarly, it is probably not a good idea to add DCHECK_NE(x, 0) in IsPowerOfTwo(), IsShiftedBit().

Patch-set: 1
Reviewer: Matteo Franchin <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1