Update patch set 12
Patch Set 12:

The last PS gets rid of the ImplicitNullCheckElimination pass. That was wrong since it didn't take into account the combinatation of dead code elimination and the side effects of the in between instructions (including environment modidication). As Nicolas suggested, the easiet correct way to remove redundant checks is to just look back at the previous instruction and record an implicit check only if that's a NullCheck. If this is done during the codegen (after register allocation) we need to take into account the additional ParallalelMoves that may be inserted between the NullCheck and the use. The parallel moves modify the  environment and their changes should be reverted. I'll address that in another CL.

Patch-set: 12