Update patch set 3
Patch Set 3: Code-Review+1

Ian, we only did this in the pause phase. There should be no reference from the alloc space to the allocation stack of the next sticky gc as the allocation stack is swapped with the live stack after this operation in the pause phase.

Patch-set: 3
Label: Code-Review=+1