Update patch set 7
Patch Set 7:

> By default (AndroidManifest/native-libraries/extractUncompressed=true),
 > or if library is compressed - PackageManager is going to extract
 > it. I am not sure where abi should come from in case of native
 > bridge.. I can modify the code to use them instead of properties.

That's still the application abi, just that in the presence of native bridge it might be different than the runtime abi. I don't think we store that anywhere now...

(to give a bit of context: in the case of native bridge the abi still comes from the list of supported abis. A system with a native bridge will list more abis (cpu abi + emulated abis). The first matched abi becomes the application abi and if that is not the same with the runtime abi then the app is emulated)

Patch-set: 7
Label: Code-Review=0