Update patch set 2
Patch Set 2:

> agampe, do we have any thoughts of using something like this in
 > run-test to prevent accidental fallback in general?

Run-tests are not at stake here. As Nicolas mentioned, the prebuild option will already fail when dex2oat goes wrong.

Intel is asking for an option where a device or emulator can be used and actual apps can be tested. I principally support this, the code here just needs to be fixed for style (long lines etc.). It should probably also get a system property read in AndroidRuntime (so that init.rc doesn't need to be changed).

Patch-set: 2