Update patch set 6
Patch Set 6:

> Sounding like an old record but: please make this a new test.

As I said before:

"""
So I wrote this test, I'm not really concerned with this refactoring. We've always been adding more to this, it really is not intended to be a simple unit test, but something to have interactions between different but overlapping areas. I was the one who requested to the new case to be added. I'm not offended that Andreas thinks my old code is painful.

I'll note we've been refactoring the old run-tests as well as we add new cases to them.
"""

I want to add other new types of stress to this test, probably daemon threads left running doing things with locks. i'm happy to keep adding more targeted regression tests for things found by this test, but I don't want to fork it just so we can make improvements.

Patch-set: 6