Update patch set 4
Patch Set 4:

> does attribute order matters?
 > 
 > It lgtm but since you're the Checker expert I leave it to you to
 > self-approve.

Yep, the order of stuff on a single line matters. I'm still trying to get by without having to hard-code too many of the output format specifics and this is one of them. We could make HGraphVisualizer sort them by name and then the order would be the same even if stuff is added/removed.

Patch-set: 4
Label: Verified=0