Update patch set 12
Patch Set 12:

I think that the problem here is an invalid test.  The test CodegenTest.ReturnMulLong gets a segmentation violation.  The reason for this is that the test is called as:
template <typename Expected>
static void Run(const InternalCodeAllocator& allocator,
                const CodeGenerator& codegen,
                bool has_result,
                Expected expected) {
  typedef Expected (*fptr)();
  CommonCompilerTest::MakeExecutable(allocator.GetMemory(), allocator.GetSize());
  fptr f = reinterpret_cast<fptr>(allocator.GetMemory());
  if (codegen.GetInstructionSet() == kThumb2) {
    // For thumb we need the bottom bit set.
    f = reinterpret_cast<fptr>(reinterpret_cast<uintptr_t>(f) + 1);
  }
  Expected result = f();

Unfortunately, the test is compiled to:
(gdb) x/19i $pc
=> 0xb5c45100:  sub    $0x1c,%esp
   0xb5c45103:  mov    %eax,(%esp)
   0xb5c45106:  movl   $0x3,0x10(%esp)
   0xb5c4510e:  movl   $0x4,0x14(%esp)
   0xb5c45116:  mov    0x10(%esp),%ecx
   0xb5c4511a:  mov    0x14(%esp),%ebx
   0xb5c4511e:  mov    0x18(%esp),%eax
   0xb5c45122:  imul   %ecx,%eax
   0xb5c45125:  imul   0x14(%esp),%ebx
   0xb5c4512a:  add    %eax,%ebx
   0xb5c4512c:  mov    %ecx,%eax
   0xb5c4512e:  mull   0x14(%esp)
   0xb5c45132:  add    %edx,%ebx
   0xb5c45134:  mov    %eax,%ecx
   0xb5c45136:  mov    %ecx,0x10(%esp)
   0xb5c4513a:  mov    %ebx,0x14(%esp)
   0xb5c4513e:  mov    0x10(%esp),%eax
   0xb5c45142:  add    $0x1c,%esp
   0xb5c45145:  ret

The test clobbers EBX, which is a callee saved register for the native C++ ABI, but not for ART (the same problem that I forgot to fix in this CL).   On return to the Run() routine, EBX is assumed to point to the GOT, which is no longer the case, and therefore we get a SEGV.

What is the proper way to handle this?

Patch-set: 12