Update patch set 1
Patch Set 1:

| Is it OK to declare them as virtual in the backend as done here? Or do we need to move it to Mir2Lir (is it some MIR analysis stuff)?

It is required in the MIR level to give the ME information about how many vectorized registers are available for it do to its work. So I think it is required in Mir2Lir actually because I think there is no way I can call it if it is in the backend directly from the MIR level.

Patch-set: 1
Reviewer: Jean Christophe Beyler <1016743@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0