Update patch set 1
Patch Set 1:

> Andreas, I want to get a clarification because your comment relates
 > to something I have been thinking about. I have noticed that unless
 > we call into boot, we always go via dex cache. For calls within
 > same dex file, we should be able to do relative calls with literal
 > offset or something in pool that can get patched for pic case. Is
 > this what you meant when you recommended adding the TODO?

Relative calls have been implemented in Quick (some of vmarko's patches), but not yet in the optimizing compiler. Issues as usual are: 1) arm32 needs islands because of small displacement, so some "linker" support, 2) it wasn't on the design agenda yet for optimizing and its assemblers. It is definitely on the road map, though, and some design input for the x86 side are always welcome (though it's the easier one, with nice 32b displacement directly in the call instruction).

The suggested TODOs here just mention that we don't have to dispatch through the method in this specific case, because we (can) exactly know the displacement (it's independent of other methods and oat layout).

Patch-set: 1