Update patch set 5
Patch Set 5: Verified+1

Works now. Tested on host. The coverage report includes a large number of directories because of all the generated sources, but there's nothing to be done about that.

btw, a lot of tests fail with coverage because they're hitting timeouts, so someone might want to up the timeout for coverage test runs.

Patch-set: 5
Label: Verified=+1