Update patch set 1
Patch Set 1:

> It's not a problem of the move resolver, it's a problem of the user/backend. I haven't checked the arm64 backend, but thje others do adjust the stack if needed.
Sorry for misleading you, it seems to be a bug in my manually created parallel move test.

> I'd rather not complicate the move resolver too much and/or have another one. If you have a simple optimization for the current one, you are going in the right direction :)
I'm not sure if x86 likes to use a scratch register to resolve the dependencies, because there is xchg instruction on x86. So I created another one... But the code won't be too complicated, the logic will be very similar with current one.

Patch-set: 1