Update patch set 1
Patch Set 1:

Hi Hiroshi, actually the motivation of reducing sweep time of ros space is to walk the coarse-grained page table of this space not generic fine-grained bitmap. The stride of page table is at least 4K bytes while the stride of walking bitmap is only 8 bytes. Also this optimization can bypass the free page holes which are usually larger than one page. When walking the page table, if it indicates the beginning of one single run, then it will find the corresponding the bitmap of this run. Using max_allocated_idx is to just avoid walking the empty bitmap of the trail of the run. Actually the overhead of calculating this index of each allocation is minor. For large object space, the optimization of reducing sweep time is to walk the metadata mem_maps_ which records total large objects. Actually large objects are usually long-living. Thus probably after a while the large objects may be distributed randomly and sparsely. It may be low efficiency to still walk the bitmap of this space. The overhead of copying this metadata in pause phase is lower, since the count of large objects are usually much smaller for real Android apps. The overhead impact of pause phase I measured is using our internal benchmark which has more large objects. Even if using this benchmark, the overhead is lower than 1 ms. Thanks.

Patch-set: 1
Label: Verified=0