Update patch set 7
Patch Set 7:

(2 comments)

> > (3 comments)
 > >
 > > The new test looks fragile because it depends on the specific
 > > wording of an error message.
 > >
 > > Can you test instead by trying to load a dex file and execute a
 > > method from it that prints something to the screen. If something
 > > was printed, the runtime must have fallen back to the dex file,
 > so
 > > the test fails. Otherwise nothing is printed, and the test
 > passes.
 > > Would you still need a special check script in this case?
 > 
 > The issue with the test is that you get the ClassNotFoundException
 > because the runtime can't find Main. So I'd have to write a
 > multiple dex solution to have Main actually catch that
 > ClassNotFoundException. That's just as complicated, so I prefer it
 > this way for now.

There are other test cases that use the multiple dex solution, right? I would prefer that, because it's more consistent with other test cases (at least the ones I have seen), and it will save effort in the future for whoever tries to change the error message and is surprised to find they broke a test.

I don't think it matters overly much. I'll leave it to you to decide.

Patch-set: 7