Update patch set 3
Patch Set 3:

The problem seems to be the assumption that register pairs are [X,X+1], which is fine for ARM, but fails miserably on X86 due to the small number of registers.

I am willing to try to enhance the RA to handle even/odd pairs. Does it make sense to add a virtual RegisterPairsEvenOdd() in the code generator to return true for ARM and false for X86?  It can even abort for other architectures.

The assumption that one register number can represent both low & high will have to change.  Is it better to split this into 2 variables, or encode Low & High parts in 2 16 bit fields in a 32 bit 'register number' and break them apart as needed?

Patch-set: 3