Update patch set 3
Patch Set 3:

> Make that 'non even/odd pairs' :-)

Does it have to fail otherwise? What's the situation that leads to not having enough registers?

I think that given the x86 situation, we can live with the not ideal strategy of even/odd. And it'd be great if we could so that the regalloc doesn't suffer from complexity again.

Patch-set: 3
Label: Verified=0