Update patch set 8
Patch Set 8:

> This might not scale as much as you'd like. It's mostly decoding
 > that is different, and you'll have to end up doing something like:
 > 
 > CodeItem code_item
 > if (dictionary_base) code_item = DictCodeItem(data);
 > else if (...) code_item = ...
 > else ...
 
We'll end up with this only if we want to support all of them in parallel which I don't think it's the case. At any point we'll only have on method enable.

 > What could look more readable is having a DictionaryDecoder,
 > ListDecoder, CompressDecoder, what not and have the read and write
 > methods on them.

The downside of this is that you'll still need to embed all the extra information in CodeInfo (e.g. extra dictionaries) even if not all methods use it.

Patch-set: 8
Reviewer: Calin Juravle <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0