Update patch set 4
Patch Set 4:

> I couldn't find where the dex files are actually being freed?

The issue was never free-ing the dex files. When closing a dex file, you free those that are not still registered in the class-linker. The leak has always been the vector itself, which is now a Java array.

 > 
 > I also don't understand the need to transform the array into a
 > std::vector?

It is simply easier than having to do the JNI array access in the places where it's used. It allows to keep most of the code as is, and isn't really much of an overhead.

Patch-set: 4