Update patch set 1
Patch Set 1:

> This makes the value loads unaligned. Should we consider a
 > different layout? How about first all values, and then the kinds.
 > That way you could actually use nibbles instead of full bytes if
 > you wanted to compress more...

We do even more unaligned accesses in https://android-review.googlesource.com/#/c/137114/.  I need to do more testing, but it seems to be working so far.  Maybe you are concerned about performance issues?  If so, I guess this should not be a huge problem, as Dex register maps are only used for debugging purpose.

Patch-set: 1
Label: Verified=0