Update patch set 1
Patch Set 1:

As I found out while working on https://android-review.googlesource.com/132771 for the X86, the current Parallel resolver doesn't handle non even/even+1 register pairs.  This might work better for that, but I found that the easiest fix was to split up the paired Moves in AddMove.

Patch-set: 1