Update patch set 2
Patch Set 2:

(10 comments)

Nice! This is definitely moving in the right way, with simplified code.

However, I note we still need a "throw location" for 2 things:
1) When throwing a new exception, we need to know the class loader of the method throwing the exception to allocate the exception object.
2) When we want to give context about a pending exception, we also print the location it was thrown from.

For 2) I think we can get it from the exception object. Can you confirm that?
For 1) it's different: we are throwing the exception so there is not stack trace yet. The class loader is the one from the top method in the stack. 

What do you think?

Patch-set: 2
Label: Code-Review=0