Update patch set 2
Patch Set 2:

(10 comments)

Thanks for the review Seb!

> For 2) I think we can get it from the exception object. Can you confirm that?

Yes. I've put in my comments that if you want that level of detail, you should probably open up a debugger. We could dump it ourselves, but it's dealing with Java code, which is not ideal to run at the point we want to dump the trace....

> For 1) it's different: we are throwing the exception so there is not stack trace yet. The class loader is the one from the top
> method in the stack. 
> What do you think?

I think you're right. I'll play with another CL to check whether we can get rid of the ThrowLocation completely.

Patch-set: 2
Label: Verified=0