Update patch set 1
Patch Set 1:

I don't see a reason to not put the arenas in the runtime. Having one arena pool seems cleaner to me. The main reason to avoid doing it for the other arenas is that compilation performance is slightly regressed, though this didn't show up on JIT benchmarks.

Patch-set: 1