Update patch set 2
Patch Set 2: Code-Review+1

LGTM...

It seems like the large increase in size is because of new operations supported for unstarted runtimes. probably worth mentioning...

and then the question is, should there be a test for exercising everything that is supposed to work to test the alternative implementations for correctness?

Patch-set: 2