Update patch set 2
Patch Set 2:

(1 comment)

CL sgtm. I just have some thoughts about "unstarted runtime" vs "transaction" I'd like to share and get your opinion.

It's getting harder to me to have a clear vision of the scope of an "unstarted runtime", especially when we deal with transactions.
A transaction can only be active at compilation time when we pre-initialize image classes. This implies the runtime is unstarted because dex2oat simply does not start it.
However unstarted runtime does not imply a transaction is active, thus the need to test whether there's one active. IIRC that may happen during the gtests, right?

My point is: should we support the same set of methods to be executed when we are not pre-initializing image classes? And for the common methods: should we behave the same in case of errors ("abort transaction" vs "throw exception" or "fail")?

Patch-set: 2