Update patch set 2
Patch Set 2:

(4 comments)

> (1 comment)
 > 
 > CL sgtm. I just have some thoughts about "unstarted runtime" vs
 > "transaction" I'd like to share and get your opinion.
 > 
 > It's getting harder to me to have a clear vision of the scope of an
 > "unstarted runtime", especially when we deal with transactions.
 > A transaction can only be active at compilation time when we
 > pre-initialize image classes. This implies the runtime is unstarted
 > because dex2oat simply does not start it.
 > However unstarted runtime does not imply a transaction is active,
 > thus the need to test whether there's one active. IIRC that may
 > happen during the gtests, right?
 > 
 > My point is: should we support the same set of methods to be
 > executed when we are not pre-initializing image classes? And for
 > the common methods: should we behave the same in case of errors
 > ("abort transaction" vs "throw exception" or "fail")?

I guess this is worth a discussion.

I kept things in line with previous behavior. My CL replaced CHECKs with AbortTransactionOrFail. This "improves" dex2oat behavior, and keeps the other abort. I added some missing checks, so we don't get segfaults in the worst cases, and abort in different locations in the better cases.

I only know of the gtests as the other thing that runs partially unstarted, and the only thing doing so is root clinits AFAIK. I don't think any tests run actual code without starting the runtime for it.

So I feel that this quite restricted behavior is acceptable.

Patch-set: 2