Update patch set 3
Patch Set 3:

> As implemented now, --fuzz *adds* fuzzing to usual testing. If I
 > continue this way, all flags will be used just as before, then
 > fuzzing will be done. It could make sense to want to make sure the
 > test passes before fuzzing it.
 > 
 > Do we want to be able to fuzz without running the actual test? If
 > so, yes, I have to check other cli options. Otherwise, what could
 > be conflictual? (really asking)
 > 
 > I'm an intern here since monday, my manager isn't currently here,
 > so I'm more asking for directions and advices than any validation
 > here.

No worries, I'm just giving my opinion...

I didn't notice from skimming the change that it would actually run the normal test. I was expecting it to be a different "run mode." The reason for that is that then we can separate it from the actual test run and not duplicate work, which helps with build bots (i.e., have different targets so you can run the test for correctness and run it for fuzzing).

In that case IMHO it's better to check for the correct command line parameters, that's what I mean by conflict. Else it's hard to notice wrong changes in the driver (the Makefile)...

Patch-set: 3
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0