Update patch set 1
Patch Set 1:

> Hey Alexandre, Checker doesn't know the target mode because it is
 > meant to work just on the HGraph. The only reason we dump the state
 > after register allocation is because we also use the C1visualizer
 > format for debugging, otherwise the output is completely
 > arch-agnostic.

I understand that. But following discussions with Nicolas, I hope that relatively soon the HGraph will not be arch-agnostic during the all compilation process. The last passes of compilation would allow for arch-specific IRs and optimizations.
 > 
 > What sort of tests do you envision here? 

Introducing the ability to have arch-specific tests now would allow me to push checker tests along with the patches introducing the arch-specific opt patches. For now it can be exercised by caller/callee saved register allocation tests.

 > Wouldn't it be better to do arch-specific testing on assembly?

I think testing the assembly will be a separate test layer. I think we can use disassembly tests to verify that an IR is compiled to assembly as expected. But as far as optimisations on the graph go, using checker tests will be more powerful and more convenient.

Patch-set: 1