Update patch set 1
Patch Set 1:

After a bit of conversation with Nicolas here is some clarification: this wasn't meant for persisting decisions over from boot compilation. It's for one compile. If a method is too complex, it shouldn't be inlined, but for some methods that decision is computed repeatedly. So this flag is meant to curb that.

Patch-set: 1