Update patch set 1
Patch Set 1:

> Interesting change. Out of curiosity, are you just trying to save
 > memory from alignment overhead?

Saving memory is actually a side effect :-) The primary reason for this change is that unaligned word load on ARM is unspecified behavior, so gcc may optimize two word loads into a single double-word load instruction that raises a SIGBUS if the data is not word aligned.

The reason it worked before is because optimizing always ensured its stack maps are word aligned. But now that there is a mix of Quick and Optimizing in the same oat file, Quick does not guarantee this, and we were getting unaligned optimizing stack map.

Patch-set: 1