Update patch set 4
Patch Set 4:

I know it's a bit unfair to ask for this, but this is a huge change. Could we figure out a way to write a test for this?

I was thinking a run-test using JNI to transition to native, and using libbacktrace or something (similar to how thread stack dumps are taken).

Patch-set: 4
Label: Verified=0