Update patch set 4
Patch Set 4:

Yes, I realize that this patch is shamefully missing any tests.  Sorry.

Since my motivation for this patch is to support the LLDB efforts, I was considering using that for testing.  That is, run a program inside LLDB and check the observed state - including single stepping instructions and reading register values.  However, before that is possible, we need to get LLDB into a good working state first.

I did not think of libbacktrace.  It seems like a good option in the short term.  Most importantly, it tells us whether we can see the whole backtrace - which is the main point of this patch.

I did my best to make sure this patch does not affect anything other than the debug sections.  After removing the code which emits the .eh_frame and .debug_* section, I compared the various core*.oat files before and after the patch, and they were bit for bit identical.  (with the small exception of Arm64 embedding __LINE__ in the assembly; that was a pain to figure out)

Patch-set: 4