Update patch set 3
Patch Set 3:

(1 comment)

> (1 comment)
 > 
 > > Shouldn't we be able to write a test for this, from the
 > relatively
 > > clear understanding what breaks?
 > 
 > There is already a smali test, 134-reg-promotion. Removing the
 > workaround (diff PS2 PS3) and seeing all the tests still pass
 > verifies this fix. Or do you want a contrived test exposing what
 > wasn't handled by the workaround?

The latter. The thing that's breaking in the bug looks like we should be able to recreate. Just call to something else that will generate objects till OOME to consistently cause a GC, if I understand it correctly. So a mix of smali and Java seems the right way to do it.

That way we could close the bug and point to the test...

Patch-set: 3