Update patch set 1
Patch Set 1: Code-Review+2

I think this is a fine change. And fixes the redness. I shamefully ignored the problem during the quick/optimizing transition, so thanks for trying to fix it.

Andreas: is readability really a problem worth thinking here? We know our current setup does not scale, so we'll need to rewrite it anyway. In the meantime, a change like this one doesn't really harm.

Patch-set: 1
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2