Update patch set 3
Patch Set 3:

(4 comments)

So with this implementation, it seems to be hard to go back, in case the dictionary is actually making things worse, right? Should we add a bit in the CodeInfo to tell if we're using the dictionary or not?

Or did we actually never encounter such a case, and we should just go unconditionally for it?

Patch-set: 3
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0