Update patch set 21
Patch Set 21:

> Dear Sebastien,
 > 
 > thank you for looking at our patch again. Locally we also rebased
 > it to the latest version on the master branch, and added an
 > additional fix to work with native frames correctly (like stepping
 > out from a java frame that is called by a native one). Should we
 > upload a new PS with this fix?
 > 
 > We also created 2 new JDWP test cases in the external/apache-harmony
 > project:
 > - we have one test case for stepping and continuing around a static
 > initialiser
 > - we have another test case for the aforementioned native frame
 > fix. This one requires one JNI function definition in C, but we
 > could not find the standard way to include such tests (we used a
 > hack in our own tree).
 > 
 > Should we submit patches for these tests? How should we include the
 > JNI function definition for the second test?

Hi,

I'd say it depends on the size of your additional change: is it only a few lines change or is it more? I was planning on rebasing this change today and ask for a final review. I'd prefer a separate CL at this point. You may want to upload a dependent CL based on this one using "git reset --hard <sha1>".

Concerning the JDWP tests, please upload CLs so I can take a look. There is no JNI support yet for the JDWP tests (I was using java.lang.reflect.Method#invoke to do the trick but it's limited). It likely requires to change the Android.mk to build a .so file and load it during the JNI tests. 
Note JDWP tests run in the CTS (starting from L release) so such change should be compatible with CTS build.

Patch-set: 21
Label: Verified=0