Update patch set 2
Patch Set 2:

(1 comment)

I'm not sure this fixes the problem entirely though. When setting the breakpoint with the JIT enabled, we'll now change all method entrypoints to the interpreter. However, it won't prevent the JIT from changing the entrypoint of a method it just compiled, even if there is a breakpoint in that method. So we'll still miss the breakpoint in that case.

In order to properly work, the entrypoint shouldn't be changed as long as the breakpoint is active. In other words: as long as the debugger needs to run in full interpreter mode (there could me more than 1 breakpoint set). 

This is probably something to manage in the instrumentation. We rely on it to remind which method we deoptimized (Instrumentation::Deoptimize) or whether we deoptimized everything (Instrumentation::DeoptimizeEverything). When we "undeoptimize", the instrumentation is supposed to restore the original entrypoint from the oat file. This should be changed to also look at the JIT's code cache I presume.

Patch-set: 2
Label: Verified=0