Update patch set 4
Patch Set 4:

That is odd.  It appears that objdump on the bot works but it produces slightly different output to my machine.  In particular, it does not contain the phrase 'Decoded dump of debug contents of section .debug_line:', I am not sure why.

Can I easily log-in to the bot machine and try to run "objdump -W -WL" on any file to see what it produces?

Alternatively, I can modify the test to dump the whole output on any failure.  What would be a good way? printf?

Patch-set: 4