Update patch set 1
Patch Set 1:

With some other optimization, the compiler might be able to know what is the behavior of AsLongConstant() and GetValue().

Also "Dereferencing a NULL Pointer" is undefined behavior in C. So both keeping the code and trigering fault in "O0" or removing the code in "O2" are correct implementation for a C/C++ compiler.

See: http://blog.llvm.org/2011/05/what-every-c-programmer-should-know.html

But we'd better fix this, as the behavior is undefined.

Patch-set: 1