Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

It's hard to verify this, as we actually don't have settings that trigger this. I tested manually by changing the default for FromVariant(), but we need to think about something for continuous host testing. The weird thing is: it seems we're not compiling with SSE4.1 as cpp flags either?!?! At least when I made the change dex2oat complained that FromVariant(default) != FromCppDefines().

Nicolas, I think it's ready to merge, but take a quick look...

Patch-set: 2
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1