Update patch set 1
Patch Set 1:

It looks like the granularity for verification errors is class level instead of method level. Is there a good reason for this?

I'm facing the issue that if we skip the entire class then we miss quite a lot of methods. I could potentially re-run verification for individual methods but that seems stupid.

Patch-set: 1
Label: Verified=0