Update patch set 1
Patch Set 1:

It turns out most of the loads hoisted between the adrp and the associated ldr are PC relative loads. (for a recent version of boot.oat). This pattern stems from the way adrp is currently used and may change in the future. Currently, however, with this patch we reduce the number of workaround applications to 2 over the whole boot.oat (it was around 50 before).

Patch-set: 1
Label: Verified=0