Update patch set 2
Patch Set 2:

> Yes, Nexus Player has SSE 4.1 support (checked on the device by a
 > colleague)

We don't set any instruction-set-features/variant right now. We should introduce silvermont as an x86 variant instead of the plain atom, and wire it up in the build.

You can't set features 'willy-nilly' in the host build. Our lowest supported host is Prescott, which does *not* have SSE4.1. Weirdly, as I mentioned, it seems we don't even take advantage of SSE4.1 on the host, as the preprocessor flags are not set on the host (looking in build, it seems we *do* set them for the target, so compiling on Fugu should actually always print a warning that default variant and runtime instruction-set-features don't match).

Patch-set: 2