Update patch set 2
Patch Set 2:

Thanks Andreas.

I hooked into the verification_results in the end to figure out for which methods we should fallback to quick. 

If we got as far as asking for the compilation but the method still has any verification failures then I delegate to quick.

PTAL.

Patch-set: 2
Label: Verified=0