Update patch set 1
Patch Set 1:

I think that would be the proper thing.

I made the wrong assumption about kArm and kThumb2.  So are they effectively an alias?

If they are an alias, why not remove one of them?

Does that mean we only support Thumb2 devices at the moment and never codegen the non-thumb Arm instructions?

If that is the case will we ever want to support non-thumb Arm?  I assume all new hardware will support Thumb2.

There are still some things in the code that confuse me:
CompiledCode::CodeDelta returns 1 for kThumb2 but not for kArm.
There is Arm32Assembler as well as Thumb2Assembler.

Patch-set: 1