Update patch set 2
Patch Set 2:

(8 comments)

Fixed all the line issues.

With this patch, we cannot build all ISAs together. This is because, all ISAs now use a common function defined in appropriate header file. So, if we try to build, it will complain that there are multiple definitions of same symbol found. 

To solve this problem, there are two approaches - 

1. Use dummy/error symbols/functions for ISAs other than TARGET_ARCH similar to the way the compiler type is being handled in this patch (If you try to invoke optimizing when only quick is built, a nullptr is returned from the error/optimizing_compiler.cc)

2. Build three libart-compiler.so - libart-compiler_x86.so, libart-compiler_arm.so and libart-compiler_mips.so). For the target link libart-compiler.so to the appropriate ISA so file. For the host, build all three ISA and make dex2oat pickup appropriate ISA specific .so after parsing the --instruction-set argument.

Which of the two approaches do you prefer? If not, is there any other suggestion?

Patch-set: 2
Label: Verified=0