Update patch set 1
Patch Set 1:

Nicolas,
Does the approach employed here look reasonable to you? I can add more restrictions if you do not agree with my approach fully. The end result of this patch is that I want it to allow us to inline getters where the object is passed by argument. Quick is able to do it because it has logic to throw exception with same offset as invoke and also because it has special logic to not care about NPE stack for synthetic accessors.

Patch-set: 1
Label: Verified=0