Update patch set 11
Patch Set 11:

With the test, this CL is again very large. Gathering opinions about splitting this into three CLs.

The first is for Quick, and basically just a rewrite/extension of the existing infrastructure. The second is for the assemblers and JNI and would contain the test in PS11. The third is the new integration into the optimizing compiler.

Patch-set: 11