Update patch set 11
Patch Set 11:

I was mostly just following the pattern of the other test (without being aware there was contention about it).  The .inc files are essentially just generated code, so it made sense to me to split them to separate file.  Having said that, the files can be easily renamed or the content can be included in the main test files. (there are three tests - Quick, JNI, Optimizing)

I agree that the split to three parts (Quick, JNI, Optimizing) would be most natural if necessary.  However, I wonder if the split is necessary.  I see it as one (although big) chunk of related work.  If things do not go perfectly and something breaks we will probably want to revert all of CFI, fix it and resubmit.  Juggling three commits might be a bit more awkward.  It also seems better to have one CL for future reference.

In a bit more practical terms, rebasing on top of master and due to proceeding renames is painful as it is.  Splitting it would triple the workload and probably prolong the review time.  (note that I have further local commits which I have to keep rebasing and which I would like to get finished)

Having said that, I know how to split this in git.  I just would like to understand why it is good idea before I spent significant time on it.

Patch-set: 11