Update patch set 1
Patch Set 1:

This is also mostly just cleanup.
I have neglected this part of DWARF when I wrote the first version.
Add it now to clean up the code and to prepare it for the future - I expect to write much more debug_info data at some point.  Having a library for it will be very useful.

Patch-set: 1
Label: Verified=0